Move: test/javascript配下のテストファイルをapp/javascript/components配下の配置に合わせて整理しました。
