+++ app/javascript/components/components/Button.js
@@ -1,13 +1,13 @@
 import React from 'react';
 
-const Button = ({ key, pxSize, pySize, color, fontColor, roundType, style, onClick, children }) => {
+const Button = ({ pxSize, pySize, color, fontColor, roundType, style, onClick, children }) => {
   // ボタンのスタイルを動的に生成
   const buttonStyle = style
     ? `py-${pySize} px-${pxSize} bg-${color} text-${fontColor} rounded-${roundType} ${style}`
     : `py-${pySize} px-${pxSize} bg-${color} text-${fontColor} rounded-${roundType}`;
 
   return (
-    <button key={key} className={buttonStyle} onClick={onClick}>
+    <button className={buttonStyle} onClick={onClick}>
       {children}
     </button>
   );
