+++ app/javascript/components/pages/Top.js
@@ -2,7 +2,6 @@ import React from 'react';
 import Text from '../components/Text';
 import Image from '../components/Image';
 import Button from '../components/Button';
-import { Link } from 'react-router-dom';
 
 const Top = () => {
   const sentences = [
@@ -25,9 +24,9 @@ const Top = () => {
           ))}
         </ul>
         <div className='mt-16 text-center'>
-          <Link to='/simulation'>
+          <a href='/simulation'>
             <Button pxSize='3' pySize='2' color='amber-vivid' fontColor='white' roundType='lg'>教育費シュミレーション</Button>
-          </Link>
+          </a>
         </div>
       </div>
       <div className='w-4/12'>
